Move TransformStatus from IdentProvider to visitors
